requirements.txt
beautifulsoup4>=4.12,<5.0
fastapi>=0.115,<1.0
httpx>=0.28,<1.0
langchain-openai>=0.3,<1.0
langchain-qdrant>=0.2,<1.0
langchain-text-splitters>=0.3,<1.0
markdown>=3.6,<4.0
pydantic-settings>=2.6,<3.0
pypdf>=5.0,<6.0
python-dotenv>=1.0,<2.0
qdrant-client>=1.11,<2.0
rank-bm25>=0.2.2,<1.0
streamlit>=1.40,<2.0
tiktoken>=0.8,<1.0
uvicorn>=0.32,<1.0
pytest>=8.0,<9.0
ruff>=0.7,<1.0
